Refactor: Request Dto to Inner class
