+++ src/main/java/com/nobrain/bookmarking/domain/user/controller/UserSignController.java
@@ -1,19 +1,16 @@
 package com.nobrain.bookmarking.domain.user.controller;
 
-import com.nobrain.bookmarking.domain.user.dto.request.UserSignInRequest;
-import com.nobrain.bookmarking.domain.user.dto.request.UserSignUpRequest;
+import com.nobrain.bookmarking.domain.user.dto.UserRequest;
 import com.nobrain.bookmarking.domain.user.service.UserSignService;
 import com.nobrain.bookmarking.global.response.model.SingleResult;
 import com.nobrain.bookmarking.global.response.service.ResponseService;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
 
-@Slf4j
 @RestController
 @RequiredArgsConstructor
 public class UserSignController {
@@ -22,7 +19,7 @@ public class UserSignController {
     private final ResponseService responseService;
 
     @PostMapping("/signup")
-    public SingleResult<Long> signUp(@Valid @RequestBody UserSignUpRequest dto) {
+    public SingleResult<Long> signUp(@Valid @RequestBody UserRequest.SignUp dto) {
         return responseService.getSingleResult(userService.signUp(dto));
     }
 
@@ -30,7 +27,7 @@ public class UserSignController {
      * @return accessToken
      */
     @PostMapping("/signin")
-    public SingleResult<String> signIn(@Valid @RequestBody UserSignInRequest dto) {
+    public SingleResult<String> signIn(@Valid @RequestBody UserRequest.SignIn dto) {
         return responseService.getSingleResult(userService.signIn(dto));
     }
 }
