--- src/main/java/com/nobrain/bookmarking/domain/user/dto/request/UserSignInRequest.java
@@ -1,11 +1,0 @@
-package com.nobrain.bookmarking.domain.user.dto.request;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.Getter;
-
-@Getter
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class UserSignInRequest {
-    private String loginId;
-    private String password;
-}
