+++ src/main/java/com/nobrain/bookmarking/domain/user/service/UserSignService.java
@@ -1,7 +1,6 @@
 package com.nobrain.bookmarking.domain.user.service;
 
-import com.nobrain.bookmarking.domain.user.dto.request.UserSignInRequest;
-import com.nobrain.bookmarking.domain.user.dto.request.UserSignUpRequest;
+import com.nobrain.bookmarking.domain.user.dto.UserRequest;
 import com.nobrain.bookmarking.domain.user.entity.User;
 import com.nobrain.bookmarking.domain.user.exception.*;
 import com.nobrain.bookmarking.domain.user.repository.UserRepository;
@@ -22,13 +21,13 @@ public class UserSignService {
     private final JwtTokenProvider jwtTokenProvider;
 
     @Transactional
-    public Long signUp(UserSignUpRequest dto) {
+    public Long signUp(UserRequest.SignUp dto) {
         validateUserDuplication(dto);
         dto.encodePassword(passwordEncoder.encode(dto.getPassword()));
         return userRepository.save(dto.toEntity()).getId();
     }
 
-    public String signIn(UserSignInRequest dto) {
+    public String signIn(UserRequest.SignIn dto) {
         User user = userService.findByLoginId(dto.getLoginId());
         if (!passwordEncoder.matches(dto.getPassword(), user.getPassword())) {
             throw new UserNotCorrectPasswordException(dto.getPassword());
@@ -37,7 +36,7 @@ public class UserSignService {
         return jwtTokenProvider.createToken(user.getUsername(), user.getRoles());
     }
 
-    private void validateUserDuplication(UserSignUpRequest dto) {
+    private void validateUserDuplication(UserRequest.SignUp dto) {
         if (userRepository.existsByEmail(dto.getEmail())) {
             throw new UserEmailDuplicationException(dto.getEmail());
         }
